requirements.txt
python-dotenv
openai
vllm
pillow
blobfile